+++ miniapp/miniapp.go
@@ -13,7 +13,7 @@ type MiniApp struct {
 // OrderInfo 后端按照下面参数构造订单参数，返回给前端
 type OrderInfo struct {
 	MchID      string `json:"mch_id"`       //Y 商户号
-	TotalFee   int    `json:"total_fee"`    //Y 金额。单位：分
+	TotalFee   int64  `json:"total_fee"`    //Y 金额。单位：分
 	OutTradeNo string `json:"out_trade_no"` //Y 用户端自主生成的订单号
 	Body       string `json:"body"`         //N 订单标题
 	Attach     string `json:"attach"`       //N 用户自定义数据，在notify的时候会原样返回
@@ -30,7 +30,7 @@ func NewMiniApp(context *context.Context) *MiniApp {
 }
 
 // GetOrderInfo 获取小程序跳转所需的参数
-func (miniApp *MiniApp) GetOrderInfo(totalFeeReq int, bodyReq, outTradeNoReq, attachReq string) (orderInfo OrderInfo, err error) {
+func (miniApp *MiniApp) GetOrderInfo(totalFeeReq int64, bodyReq, outTradeNoReq, attachReq string) (orderInfo OrderInfo, err error) {
 	orderInfo.MchID = miniApp.MchID
 	orderInfo.TotalFee = totalFeeReq
 	orderInfo.OutTradeNo = outTradeNoReq
